app/requirements.txt
pytest==4.6.5
python-geoip==1.2
python-geoip-geolite2==2015.303
user-agents==2.0
